Refactor: Split string_types.rs into focused test modules
Split the monolithic string_types.rs (1,042 lines) into organized modules:

- char_types.rs: CHAR, CHARACTER, NCHAR, NATIONAL CHARACTER tests (13 tests)
- varchar_types.rs: VARCHAR, CHARACTER VARYING, NCHAR VARYING, NVARCHAR,
  NATIONAL VARCHAR tests (26 tests)
- binary_types.rs: BINARY, VARBINARY tests (7 tests)

Benefits:
- Improved test organization and discoverability
- Faster incremental compilation
- Easier to locate specific type tests
- Better maintainability for adding new variants

All 46 tests pass successfully.

Closes #2099
