+++ crates/vibesql-parser/src/tests/create_table/string_types/mod.rs
@@ -1,0 +1,5 @@
+// String type parsing tests organized by category
+
+mod char_types;
+mod varchar_types;
+mod binary_types;
